backend/dj_ng_api_django/api/serializers.py
from rest_framework import serializers
from api.models import Item


class ItemSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Item
        fields = ('id', 'name', 'desc')
