backend/requirements.txt
boto3==1.12.47
botocore==1.15.47
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.12
django-admin-rangefilter==0.5.4
django-allauth==0.41.0
django-anymail==7.1.0
django-autoslug==1.9.7
django-braces==1.14.0
django-ckeditor-filebrowser-filer==0.3.0
django-classy-tags==1.0.0
django-cms==3.7.2
django-cors-headers==3.2.1
django-crispy-forms==1.9.0
django-environ==0.4.5
django-filer==1.7.0
django-floppyforms==1.8.0
django-formtools==2.2
django-impersonate==1.5
django-import-export==2.0.2
django-js-asset==1.2.2
django-json-widget==1.0.1
django-model-utils==4.0.0
django-mptt==0.11.0
django-polymorphic==2.1.2
django-redis==4.11.0
django-resized==0.3.10
django-sekizai==1.1.0
django-static-sitemaps==4.6.0
django-storages==1.9.1
django-tinymce==3.0.2
django-treebeard==4.3.1
djangocms-admin-style==1.5.0
djangocms-text-ckeditor==3.9.0
djangorestframework==3.11.0
docutils==0.15.2
easy-thumbnails==2.7
et-xmlfile==1.0.1
future==0.18.2
gunicorn==20.0.4
html5lib==1.0.1
idna==2.9
jdcal==1.4.1
jmespath==0.9.5
markdown2==2.3.8
MarkupPy==1.14
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.2
psycopg2==2.8.5
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
redis==3.4.1
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
tablib==1.1.0
Unidecode==1.1.1
urllib3==1.25.9
webencodings==0.5.1
xlrd==1.2.0
xlwt==1.3.0
